// src/core/enums/perdida-type.enum.ts
export enum PerdidaType{
    THIEFS = 1,
    ROBBERY = 2,
    VANDALISM = 3,
    FRAUD = 4,
    KIDNAPPING = 5,
    TRAFFIC= 6, 
    ACCIDENTS = 7,
    POTHOLE = 8, 
    BROKEN_TRAFFIC_LIGHT = 9,
    FIRE = 10,
    EXPLOSION = 11, 
    GAS_LEAK = 12,
    COLLAPSE = 13, 
    FLOOD = 14, 
    EARTHQUAKE = 15,
    LANDSLIDE = 16,
    POWER_OUTAGE = 17,
    WATER_LEAK = 18, 
    OPEN_MANHOLE = 19, 
    STREET_LIGHT_OUT = 20,
    MEDICAL_EMERGENCY= 21,
    POLLUTION = 22, 
    OTHER = 23,
}